+++ backend/mora/service/employee.py
@@ -91,6 +91,7 @@ class EmployeeRequestHandler(handlers.RequestHandler):
         cpr = util.checked_get(req, mapping.CPR_NO, "", required=False)
         userid = util.get_uuid(req, required=False) or str(uuid.uuid4())
         bvn = util.checked_get(req, mapping.USER_KEY, userid)
+        seniority = req.get(mapping.SENIORITY, None)
 
         try:
             valid_from = \
@@ -112,6 +113,7 @@ class EmployeeRequestHandler(handlers.RequestHandler):
             efternavn=surname,
             kaldenavn_fornavn=nickname_givenname,
             kaldenavn_efternavn=nickname_surname,
+            seniority=seniority,
             brugervendtnoegle=bvn,
             tilhoerer=org_uuid,
             cpr=cpr,
@@ -193,10 +195,18 @@ class EmployeeRequestHandler(handlers.RequestHandler):
 
         nickname_givenname, nickname_surname = self._handle_nickname(data)
 
+        seniority = data.get(mapping.SENIORITY, None)
+
+        # clear rather than skip if exists, but value is None
+        if seniority is None and mapping.SENIORITY in data:
+            seniority = ''
+
         if nickname_givenname is not None:
             changed_extended_props['kaldenavn_fornavn'] = nickname_givenname
         if nickname_surname is not None:
             changed_extended_props['kaldenavn_efternavn'] = nickname_surname
+        if seniority is not None:
+            changed_extended_props['seniority'] = seniority
 
         if mapping.INTEGRATION_DATA in data:
             changed_props['integrationsdata'] = common.stable_json_dumps(
@@ -231,7 +241,6 @@ class EmployeeRequestHandler(handlers.RequestHandler):
             mapping.EMPLOYEE_FIELDS.difference({x[0] for x in update_fields}))
         payload = common.ensure_bounds(new_from, new_to, bounds_fields,
                                        original, payload)
-
         self.payload = payload
         self.uuid = userid
         self.trigger_dict[Trigger.EMPLOYEE_UUID] = userid
@@ -331,6 +340,7 @@ async def get_one_employee(c: lora.Connector, userid,
     efternavn = extensions.get('efternavn', '')
     kaldenavn_fornavn = extensions.get('kaldenavn_fornavn', '')
     kaldenavn_efternavn = extensions.get('kaldenavn_efternavn', '')
+    seniority = extensions.get(mapping.SENIORITY, '')
 
     r = {
         mapping.GIVENNAME: fornavn,
@@ -340,6 +350,7 @@ async def get_one_employee(c: lora.Connector, userid,
         mapping.NICKNAME_SURNAME: kaldenavn_efternavn,
         mapping.NICKNAME: " ".join((kaldenavn_fornavn, kaldenavn_efternavn)).strip(),
         mapping.UUID: userid,
+        mapping.SENIORITY: seniority,
     }
 
     if details is EmployeeDetails.FULL: